Finished database schema migration
