+++ src/database/decision-db.ts
@@ -36,14 +36,19 @@ export class DecisionInfo {
 export class DecisionEntry {
     @prop({ required: true })
     public _id: string;
+
     @prop({ required: true })
     public userId: string;
+
     @prop({ required: true })
     public wave: number;
+
     @prop({ required: true })
     public reviewer: string;
+
     @prop({ required: true })
     public timestamp: number;
+
     @prop({ required: true })
     public decision: DecisionStatus;
 }
