--- src/services/user/user-schemas.ts
@@ -1,15 +1,0 @@
-import { Document, ObjectId, WithId } from "mongodb";
-
-// User object document schema
-export interface UserSchema extends WithId<Document> {
-    _id: ObjectId;
-    id: string;
-    firstName: string;
-    lastName: string;
-    email: string;
-}
-
-// Collections within the user database
-export enum UserDB {
-    INFO = "info",
-}
